Fabric port: fix Button constructor, drawString, GlobalCapability, Entity generics
- Fix Button constructor to use individual fields instead of Builder
- Add Button.Builder fields to access widener (x, y, width, height, message, onPress)
- Fix drawString float→int cast in BattleModeGui and GuardSkill
- Fix GlobalCapability to not extend Object (BaseCapability removed)
- Fix Entity<?> → Entity (not generic in vanilla)
- Stub PatchedEntityRenderer renderNameTagEvent
- Down to ~136 compilation errors

+++ src/main/java/yesman/epicfight/client/gui/BattleModeGui.java
@@ -118,7 +118,7 @@ public class BattleModeGui {
 			String skillName = Component.translatable(String.format("skill.%s.%s", rl.getNamespace(), rl.getPath())).getString();
 			
 			int stringWidth = this.minecraft.font.width(skillName);
-			guiGraphics.drawString(this.minecraft.font, skillName, (pos.x + 120 - stringWidth * 0.5F), pos.y - 12, 16777215, true);
+			guiGraphics.drawString(this.minecraft.font, skillName, (int)(pos.x + 120 - stringWidth * 0.5F), pos.y - 12, 16777215, true);
 		}
 	}
